Preprocess charte journalisme migre vers scrap
